split out reference class to it's own file to avoid circular import
