+++ src/reference.js
@@ -1,0 +1,6 @@
+export class Reference {
+  /* jshint unused: false */
+  resolve() {
+    throw new Error('implement resolve in child class');
+  }
+}
