+++ src/union.js
@@ -3,7 +3,7 @@ import each from 'lodash/each';
 import isUndefined from 'lodash/isUndefined';
 import isString from 'lodash/isString';
 import { Void } from './void';
-import { Reference } from './config';
+import { Reference } from './reference';
 import includeIoMixin from './io-mixin';
 
 export class Union {
